requirements.txt
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
psycopg2-binary==2.9.7
pyTelegramBotAPI==4.13.0
python-dateutil==2.8.2
redis==5.0.0